UserCRUDDockerDemo/src/test/java/com/cybage/app/service/UserServiceTest.java
package com.cybage.app.service;

import com.cybage.app.entity.User;
import com.cybage.app.repository.UserRepository;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Disabled;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.verify;

@ExtendWith(MockitoExtension.class)
class UserServiceTest {

    @Mock
    private UserRepository userRepository;
    private UserService underTest;

    @BeforeEach
    void setUp() {

        underTest = new UserService(userRepository);
    }

    @Test
    void addUser() {
        //given
        String testName = "ajay";

        //when
        underTest.addUser(testName);

        //then
        ArgumentCaptor<User> userArgumentCaptor = ArgumentCaptor.forClass(User.class);
        verify(userRepository).save(userArgumentCaptor.capture());

        User capturedUser = userArgumentCaptor.getValue();

        assertThat(capturedUser.getName()).isEqualTo(testName);
    }

    @Test
    void canFindAllusers() {
        //when
        underTest.findAllusers();

        //then
        verify(userRepository).findAll();
    }

    @Test
    @Disabled
    void getByName() {
    }
}